requirements.txt
Django>=4.1.5,<4.2
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.2.2,<5.3
psycopg2>=2.9.5,<2.10
django-filter>=22.1,<22.2
